feat: extract the agent tool surface into a dedicated nokv-agent crate
Lift the seven-verb agent tool layer (tool definitions, the execute_agent_tool
dispatcher, argument validation, result shaping, the AgentNamespace trait, the
embedded impl for NoKvFs, and a transport-neutral AgentError) out of
nokv-client into a new nokv-agent crate. It depends only on nokv-meta,
nokv-object, and nokv-types -- not on nokv-client, nokv-protocol, or
nokv-control -- so the embedded in-process agent path no longer drags the RPC
stack into consumers like the benchmark harness and future SDK/MCP servers.

nokv-client keeps the remote trait impls (MetadataClient, NoKvFsClient) and
From<ClientError> for AgentError, and re-exports the surface, so existing
nokv_client call sites compile unchanged. The embedded impl lives in nokv-agent
(orphan-rule legal), so nokv-meta gains no dependency on nokv-agent and the
crate graph stays acyclic.

Bump holt 0.7.1 -> 0.7.3, which fixes read-amplification growth of the local
metadata store (blobs.dat stays flat across read-only workloads instead of
growing unbounded).

Add docs/development/nokv-agent.md (contributor handbook) and update the README
crate list.
