+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/nokv-control",
     "crates/nokv-object",
     "crates/nokv-client",
+    "crates/nokv-agent",
     "crates/nokv-fuse",
     "crates/nokv-server",
     "crates/nokv",
@@ -29,13 +30,14 @@ nokv-meta = { path = "crates/nokv-meta", version = "0.1.0" }
 nokv-control = { path = "crates/nokv-control", version = "0.1.0" }
 nokv-object = { path = "crates/nokv-object", version = "0.1.0" }
 nokv-client = { path = "crates/nokv-client", version = "0.1.0" }
+nokv-agent = { path = "crates/nokv-agent", version = "0.1.0" }
 nokv-fuse = { path = "crates/nokv-fuse", version = "0.1.0" }
 nokv-server = { path = "crates/nokv-server", version = "0.1.0" }
 nokv = { path = "crates/nokv", version = "0.1.0" }
 nokv-python = { path = "crates/nokv-python", version = "0.1.0" }
 nokv-bench = { path = "bench", version = "0.1.0" }
 fuser = { version = "0.17.0", default-features = false }
-holt = { version = "0.7.1", default-features = false }
+holt = { version = "0.7.3", default-features = false }
 etcd-client = { version = "0.19.0", default-features = false }
 libc = "0.2"
 opendal = { version = "0.57.0", default-features = false, features = ["blocking", "executors-tokio", "reqwest-rustls-tls", "services-s3"] }
